error in payment session

+++ controllers/cartControllers.js
@@ -1,19 +1,14 @@
-const { result } = require('lodash');
 const _ = require('lodash');
-const {CartItem} = require('../models/cartItem');
+const { CartItem } = require('../models/cartItem');
 
 module.exports.createCartItem = async (req, res) => {
-    let {price, product } = _.pick(req.body, ["price", 'product']);
+    let { price, product } = _.pick(req.body, ["price", "product"]);
     const item = await CartItem.findOne({
         user: req.user._id,
         product: product,
     });
-    if (item) return res.status(400).send("Item Exists in Cart!!");
-    let cartItem = new CartItem({
-        price: price,
-        product: product,
-        user: req.user._id
-    });
+    if (item) return res.status(400).send("Item already exists in Cart!");
+    let cartItem = new CartItem({ price: price, product: product, user: req.user._id });
     const result = await cartItem.save();
     res.status(201).send({
         message: "Added to cart successfully!",
@@ -27,25 +22,19 @@ module.exports.getCartItem = async (req, res) => {
     })
         .populate('product', 'name')
         .populate('user', 'name')
-    res.status(201).send(cartItems);
+    return res.status(200).send(cartItems);
 }
 
 module.exports.updateCartItem = async (req, res) => {
-    const {_id, count } = _.pick(req.body, ["count", '_id']);
+    const { _id, count } = _.pick(req.body, ["count", "_id"]);
     userId = req.user._id;
-    await CartItem.updateOne({
-        _id: _id, 
-        user: userId
-    },{ count: count });
-    res.status(200).send("Item Updated!!");
+    await CartItem.updateOne({ _id: _id, user: userId }, { count: count });
+    return res.status(200).send("Item updated!!");
 }
 
 module.exports.deleteCartItem = async (req, res) => {
     const _id = req.params.id;
     userId = req.user._id;
-    await CartItem.deleteOne({
-        _id: _id, 
-        user: userId, 
-    });
-    res.status(201).send("deleted");
+    await CartItem.deleteOne({ _id: _id, user: userId });
+    return res.status(200).send("Deleted!");
 }
